+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed issue of Angular components not being disposed (#561)
 - Improved loading of satellite dependencies in `piral-blazor`
 - Added `piral run-emulator` command to `piral-cli`
+- Added support for capabilities to `piral-blazor`
 
 ## 0.15.1 (November 25, 2022)